auth

+++ src/api/songs/handler.js
@@ -16,8 +16,9 @@ class SongsHandler {
         try {
             this._validator.validateSongPayload(request.payload)
             const { title = 'untitled', year, performer, genre, duration } = request.payload
+            const { id: credentialId } = request.auth.credentials
 
-            const songId = await this._service.addSong({ title, year, performer, genre, duration })
+            const songId = await this._service.addSong({ title, year, performer, genre, duration, owner: credentialId })
 
             const response = h.response({
                 status: 'success',
@@ -49,8 +50,9 @@ class SongsHandler {
         }
     }
 
-    async getSongsHandler (request, h) {
-        const songs = await this._service.getSongs()
+    async getSongsHandler (request) {
+        const { id: credentialId } = request.auth.credentials
+        const songs = await this._service.getSongs(credentialId)
         return {
             status: 'success',
             data: {
@@ -62,7 +64,9 @@ class SongsHandler {
     async getSongByIdHandler (request, h) {
         try {
             const { id } = request.params
+            const { id: credentialId } = request.auth.credentials
 
+            await this._service.verifysongOwner(id, credentialId)
             const song = await this._service.getSongById(id)
             return {
                 status: 'success',
@@ -93,7 +97,9 @@ class SongsHandler {
         try {
             this._validator.validateSongPayload(request.payload)
             const { id } = request.params
+            const { id: credentialId } = request.auth.credentials
 
+            await this._service.verifysongOwner(id, credentialId)
             await this._service.editSongById(id, request.payload)
 
             return {
@@ -122,7 +128,9 @@ class SongsHandler {
     async deleteSongByIdHandler (request, h) {
         try {
             const { id } = request.params
+            const { id: credentialId } = request.auth.credentials
 
+            await this._service.verifysongOwner(id, credentialId)
             await this._service.deleteSongById(id)
 
             return {
